seperates router functions, shortens Summary type fields
