+++ faker/fakedata/20180101.json
@@ -1,13 +1,9 @@
 {
   "Summary": {
     "Beginning": "5000.00",
-    "Ending": "4432.04",
-    "Deposits": "2528.94",
-    "Withdrawals": "3096.90",
-    "Debit": "",
-    "Electronic": "",
-    "Checks": "",
-    "Fees": ""
+    "Ending": "4442.27",
+    "Deposits": "2525.45",
+    "Withdrawals": "3083.18"
   },
   "Deposits": [
     {
@@ -17,7 +13,7 @@
     },
     {
       "Date": "02/22",
-      "Amount": "28.94",
+      "Amount": "25.45",
       "Description": "Venmo\tCashout\tPPD ID: 1234567890"
     }
   ],
@@ -29,7 +25,7 @@
     },
     {
       "Date": "02/12",
-      "Amount": "36.66",
+      "Amount": "39.34",
       "Description": "Card Purchase\t02/12 Dive Bar Restaurant LLC Card 1234"
     },
     {
@@ -49,7 +45,7 @@
     },
     {
       "Date": "02/24",
-      "Amount": "42.26",
+      "Amount": "25.86",
       "Description": "Card Purchase\t02/24 Dive Bar Restaurant LLC Card 1234"
     }
   ],
