+++ faker/fakedata/20180201.json
@@ -1,0 +1,58 @@
+{
+  "Summary": {
+    "Beginning": "5000.00",
+    "Ending": "4378.89",
+    "Deposits": "2512.99",
+    "Withdrawals": "3134.10"
+  },
+  "Deposits": [
+    {
+      "Date": "03/20",
+      "Amount": "2500.00",
+      "Description": "Job LLC"
+    },
+    {
+      "Date": "03/22",
+      "Amount": "12.99",
+      "Description": "Venmo\tCashout\tPPD ID: 1234567890"
+    }
+  ],
+  "Withdrawals": [
+    {
+      "Date": "03/01",
+      "Amount": "1500.00",
+      "Description": "Recurring Card Purchase\t03/01 Landlord LLC Card 1234"
+    },
+    {
+      "Date": "03/12",
+      "Amount": "26.99",
+      "Description": "Card Purchase\t03/12 Dive Bar Restaurant LLC Card 1234"
+    },
+    {
+      "Date": "03/14",
+      "Amount": "7.99",
+      "Description": "Card Purchase\t03/14 Netflix.Com Netflix.Com CA Card 1234"
+    },
+    {
+      "Date": "03/15",
+      "Amount": "1500.00",
+      "Description": "Recurring Card Purchase\t03/15 Landlord LLC Card 1234"
+    },
+    {
+      "Date": "03/23",
+      "Amount": "9.99",
+      "Description": "Recurring Card Purchase\t03/23 Spotify USA Card 1234"
+    },
+    {
+      "Date": "03/24",
+      "Amount": "20.96",
+      "Description": "Card Purchase\t03/24 Dive Bar Restaurant LLC Card 1234"
+    },
+    {
+      "Date": "03/30",
+      "Amount": "68.17",
+      "Description": "Card Purchase\t03/30 Gourmet Restaurant LLC Card 1234"
+    }
+  ],
+  "Checks": null
+}
